integrate all metrics and clean up codebase
