+++ peer-app/Dockerfile
@@ -2,6 +2,7 @@ FROM python:3.8-slim-buster
 
 RUN apt update
 RUN apt install -y iperf3
+RUN apt install -y gcc python3-dev
 
 WORKDIR /app
