+++ peer-app/requirements.txt
@@ -28,6 +28,7 @@ docopt==0.6.2
 iperf3==0.1.11
 kubernetes==25.3.0
 oauthlib==3.2.2
+psutil==5.9.4
 python-dateutil==2.8.2
 PyYAML==6.0
 requests==2.28.1
